Added addActivity method.
The execute method should only be executed on command context. Otherwise passing information through it requires a lot of hacks and workarounds.

Instead addActivity should be used on other contexts (J! plugins as example) by passing the data of the activity directly.

Additional data grabbed from request (as IP) should be added in this method.

re #60